作業の保存 image_objectクラスを作成
